app/requirement.txt
appnope==0.1.3
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.4
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
